+++ grater/internal/mockcontainer/mockdockercontroller.go
@@ -1,8 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-// Package dockercontroller provides a mock implementation of the DockerController interface.
-package dockercontroller
+// Package mockcontainer provides a mock implementation of the DockerController interface.
+package mockcontainer
 
 import (
 	"github.com/moby/moby/client"
@@ -10,22 +10,22 @@ import (
 	"go.opentelemetry.io/build-tools/grater/internal/container"
 )
 
-// MockDockerController is a mock implementation of the DockerController interface.
-type MockDockerController struct {
+// MockDockerContainer is a mock implementation of the DockerContainer interface.
+type MockDockerContainer struct {
 	CreateVolumeMock   func(string) (func(), error)
 	UseContainerMock   func(string, []string) (string, func(), error)
 	ExecuteCommandMock func(string, []string) (string, client.ExecInspectResult, error)
 }
 
-var _ container.Container = (*MockDockerController)(nil)
+var _ container.Container = (*MockDockerContainer)(nil)
 
-// NewMockDockerController creates a new instance of MockDockerController.
-func NewMockDockerController() *MockDockerController {
-	return &MockDockerController{}
+// NewMockDockerContainer creates a new instance of MockDockerContainer.
+func NewMockDockerContainer() *MockDockerContainer {
+	return &MockDockerContainer{}
 }
 
 // CreateVolume creates a mock instance of Volume.
-func (m *MockDockerController) CreateVolume(volumeName string) (func(), error) {
+func (m *MockDockerContainer) CreateVolume(volumeName string) (func(), error) {
 	if m.CreateVolumeMock != nil {
 		return m.CreateVolumeMock(volumeName)
 	}
@@ -33,7 +33,7 @@ func (m *MockDockerController) CreateVolume(volumeName string) (func(), error) {
 }
 
 // UseContainer creates a mock instance of Container.
-func (m *MockDockerController) UseContainer(imageName string, volumeNames []string) (string, func(), error) {
+func (m *MockDockerContainer) UseContainer(imageName string, volumeNames []string) (string, func(), error) {
 	if m.UseContainerMock != nil {
 		return m.UseContainerMock(imageName, volumeNames)
 	}
@@ -41,7 +41,7 @@ func (m *MockDockerController) UseContainer(imageName string, volumeNames []stri
 }
 
 // ExecuteCommand creates a mock instance of executing a command.
-func (m *MockDockerController) ExecuteCommand(containerID string, cmd []string) (string, client.ExecInspectResult, error) {
+func (m *MockDockerContainer) ExecuteCommand(containerID string, cmd []string) (string, client.ExecInspectResult, error) {
 	if m.ExecuteCommandMock != nil {
 		return m.ExecuteCommandMock(containerID, cmd)
 	}
